fix: use addRoute refresh page 404
